feat: implement receiving messages
Server now parses url in order to
use it when payloads are downloaded.
Messages are received in two steps
, first is when server notifies client
over SSE stream about new message arrival.
Second is when client downloads payload
of that message using path of url provided
in the message-received event.
Test server is now using echo for easier
adding new handlers in runtime for registering
payloads